Добавлено разделение репозиториев и эндпоинта на несколько файлов с реализацией конкретной логики в каждом файле
