+++ internal/app/endpoint/queue/get.go
@@ -1,0 +1,9 @@
+package queue
+
+import "net/http"
+
+func getFromQueue(queueSvc queueSvc) http.HandlerFunc {
+	return func(w http.ResponseWriter, request *http.Request) {
+		// TODO: parse req and call queueSvc.Get(...)
+	}
+}
