+++ internal/app/endpoint/queue/put.go
@@ -1,0 +1,9 @@
+package queue
+
+import "net/http"
+
+func putToQueue(queueSvc queueSvc) http.HandlerFunc {
+	return func(w http.ResponseWriter, request *http.Request) {
+		// TODO: parse req and call queueSvc.Put(...)
+	}
+}
