--- internal/pkg/repository/file.go
@@ -1,23 +1,0 @@
-package repository
-
-import (
-	"github.com/phpCoder88/web-broker/internal/pkg/model"
-)
-
-type FileRepo struct {
-	fileName string
-}
-
-func NewFileRepo(fileName string) *FileRepo {
-	return &FileRepo{fileName: fileName}
-}
-
-func (fr *FileRepo) Get(getReq string) (string, error) {
-	// TODO: impl
-	return "", nil
-}
-
-func (fr *FileRepo) Put(putReq *model.PutValue) error {
-	// TODO: impl
-	return nil
-}
