+++ internal/pkg/repository/file/file.go
@@ -1,0 +1,9 @@
+package file
+
+type Repo struct {
+	fileName string
+}
+
+func NewRepo(fileName string) *Repo {
+	return &Repo{fileName: fileName}
+}
