+++ internal/pkg/repository/file/get.go
@@ -1,0 +1,8 @@
+package file
+
+import "github.com/phpCoder88/web-broker/internal/pkg/model"
+
+func (fr *Repo) Put(putReq *model.PutValue) error {
+	// TODO: impl
+	return nil
+}
