+++ internal/pkg/repository/file/put.go
@@ -1,0 +1,6 @@
+package file
+
+func (fr *Repo) Get(getReq string) (string, error) {
+	// TODO: impl
+	return "", nil
+}
