+++ internal/pkg/repository/memory/get.go
@@ -1,0 +1,6 @@
+package memory
+
+func (mr *Repo) Get(getReq string) (string, error) {
+	// TODO: impl
+	return "", nil
+}
