+++ internal/pkg/repository/memory/memory.go
@@ -1,0 +1,7 @@
+package memory
+
+type Repo struct{}
+
+func NewRepo() *Repo {
+	return &Repo{}
+}
