+++ internal/pkg/repository/memory/put.go
@@ -1,0 +1,8 @@
+package memory
+
+import "github.com/phpCoder88/web-broker/internal/pkg/model"
+
+func (mr *Repo) Put(putReq *model.PutValue) error {
+	// TODO: impl
+	return nil
+}
