+++ internal/pkg/repository/postgres/get.go
@@ -1,0 +1,6 @@
+package postgres
+
+func (pgr *Repo) Get(getReq string) (string, error) {
+	// TODO: impl
+	return "", nil
+}
