+++ internal/pkg/repository/postgres/postgres.go
@@ -1,0 +1,7 @@
+package postgres
+
+type Repo struct{}
+
+func NewRepo() *Repo {
+	return &Repo{}
+}
