+++ internal/pkg/repository/postgres/put.go
@@ -1,0 +1,8 @@
+package postgres
+
+import "github.com/phpCoder88/web-broker/internal/pkg/model"
+
+func (pgr *Repo) Put(putReq *model.PutValue) error {
+	// TODO: impl
+	return nil
+}
